move wrapped resource from dx12 swapchain
