+++ src/Features/Upscaling/DX12SwapChain.h
@@ -11,18 +11,7 @@
 
 #include <directx/d3dx12.h>
 
-class WrappedResource
-{
-public:
-	WrappedResource(D3D11_TEXTURE2D_DESC a_texDesc, ID3D11Device5* a_d3d11Device, ID3D12Device* a_d3d12Device);
-	~WrappedResource();
-
-	ID3D11Texture2D* resource11 = nullptr;
-	ID3D11ShaderResourceView* srv = nullptr;
-	ID3D11UnorderedAccessView* uav = nullptr;
-	ID3D11RenderTargetView* rtv = nullptr;
-	winrt::com_ptr<ID3D12Resource> resource;
-};
+#include "Features/DX12Interop/WrappedResource.h"
 
 struct DXGISwapChainProxy : IDXGISwapChain
 {
